Fix unit tests for title
